refactor: move AsyncExecutionHandle to models layer
Remove clients → agents upward dependency by moving AsyncExecutionHandle
from vibe3.agents.backends.async_launcher to vibe3.models.execution_handle.

Changes:
- Create src/vibe3/models/execution_handle.py with AsyncExecutionHandle dataclass
- Update async_launcher.py to import and re-export from models (backward compatible)
- Update clients/protocols.py to import from models instead of agents

This resolves the architecture violation where clients layer imported from
agents layer (upward dependency). AsyncExecutionHandle is pure data model
and belongs in the models layer.

Fixes #1252
