+++ src/vibe3/clients/protocols.py
@@ -3,7 +3,7 @@
 from pathlib import Path
 from typing import Any, Protocol
 
-from vibe3.agents.backends.async_launcher import AsyncExecutionHandle
+from vibe3.models.execution_handle import AsyncExecutionHandle
 from vibe3.models.pr import CreatePRRequest, PRResponse, UpdatePRRequest
 from vibe3.models.review_runner import AgentOptions, AgentResult
